+++ ProjectManagement.Shared/DTO/ProjectAppStorage/Service/ITasksUserComputationService.cs
@@ -60,7 +60,7 @@ public sealed class TasksUserComputationServiceWasm : ITasksUserComputationServi
                 foreach (var item in res.CostRole)
                     if (item.Min <= res.Data.Quantity.Value && item.Max >= res.Data.Quantity.Value)
                     {
-                        res.Data.Cost = item.Value.Value;
+                        res.Data.Cost = (decimal)item.Value.Value;
                         break;
                     }
         }
